% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_summary.R
\name{rd_get_results}
\alias{rd_get_results}
\title{Get target number of events}
\usage{
rd_get_results(
  rst_summary,
  target_event = NULL,
  multi = NULL,
  power_by = NULL,
  power_level = 0.9
)
}
\description{
Get target number of events
}
