% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_interim.R
\name{rd_interim_test}
\alias{rd_interim_test}
\title{Get interim test results}
\usage{
rd_interim_test(
  rst_rejection,
  target_event,
  inter_looks,
  hyp_test,
  n_cores = 10,
  ...
)
}
\description{
Get interim test results
}
