% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_simu.R
\name{rd_pts_all}
\alias{rd_pts_all}
\title{Simulate Survival time for all}
\usage{
rd_pts_all(
  ve_trt,
  ir_placebo_1yr,
  p_evaluable = 1,
  annual_dropout = 0.025,
  annual_enroll = 6000,
  n_tot = 15000,
  max_followup_yr = 1,
  seed = NULL,
  ...
)
}
\description{
Simulate Survival time for all
}
