% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_tools.R
\name{rd_pts_single}
\alias{rd_pts_single}
\title{Simulate Survival time for a single arm}
\usage{
rd_pts_single(n, ve, lambda_placebo, lambda_censor, p_evaluable, max_followup)
}
\description{
Simulate Survival time for a single arm
}
