% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_shiny.R
\name{rd_shiny}
\alias{rd_shiny}
\title{Run Web-Based application}
\usage{
rd_shiny()
}
\description{
Call Shiny to run \code{ctdesign} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
