% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_simu.R
\name{rd_simu_all}
\alias{rd_simu_all}
\title{Simulate all}
\usage{
rd_simu_all(
  reps = 1:2000,
  scenario = 1,
  ir_placebo_1yr = 0.04,
  n_cores = 5,
  update_progress = NULL,
  ...,
  seed = NULL
)
}
\description{
Simulate all
}
