% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_simu_surv.R
\name{rd_simu_pwexp}
\alias{rd_simu_pwexp}
\title{Simulate from piecewise constant exponential}
\usage{
rd_simu_pwexp(hazards, offset = 0)
}
\description{
Simulate from piecewise constant exponential
}
\examples{
rd_sim_pwexp(hazards = c("0.5" = 0.045, "Inf" = 0.025), offset = 0)

}
