% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_tools.R
\name{rd_pts_all}
\alias{rd_pts_all}
\title{Simulate Survival time for all}
\usage{
rd_pts_all(
  ve_trt,
  lambda_placebo,
  p_evaluable = 0.8,
  n_tot = 15000,
  annual_dropout = 0.05,
  annual_enroll = 7500,
  max_followup = 1,
  seed = NULL,
  ...
)
}
\description{
Simulate Survival time for all
}
